+++ fcnui_base/lib/src/src.dart
@@ -1,10 +1,3 @@
-export 'package:flutter_redux/flutter_redux.dart';
-export 'package:redux/redux.dart';
-
-export 'package:flex_color_scheme/flex_color_scheme.dart';
-
-export 'package:equatable/equatable.dart';
-
 export 'store/store.dart';
 export 'vm_providers/vm_providers.dart';
 export 'dependency_injection/dependency_injection.dart';
@@ -12,3 +5,13 @@ export 'dependency_injection/dependency_injection.dart';
 export 'helpers/helpers.dart';
 
 export 'utils/utils.dart';
+
+export 'package:flutter_redux/flutter_redux.dart';
+export 'package:redux/redux.dart';
+
+export 'package:flex_color_scheme/flex_color_scheme.dart';
+
+export 'package:equatable/equatable.dart';
+
+export 'package:flutter_form_builder/flutter_form_builder.dart';
+export 'package:form_builder_validators/form_builder_validators.dart';